services/trade_to_ohlc/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Ramane23 <contactsouley@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
quixstreams = "^2.8.0"
fire = "^0.6.0"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
